+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - PR #871 CSV Reader: Add support for NA values, including user specified strings
 - PR #867 CSV Reader: Add support for ignoring blank lines and comment lines
 - PR #895 Add Series groupby
+- PR #898 Add DataFrame.groupby(level=0) support
 
 ## Improvements